requirements.txt
schedule~=0.6.0
python-dotenv~=0.15.0
pytelegrambotapi~=0.3.0